+++ changelog.md
@@ -7,6 +7,7 @@
 - Added "Infinite Beam Ammo Fix", preventing the ammo for beam weapons from decreasing when under the effects of Magazine Rack
 - Added "Caped Bullet Kin Fix", preventing Caped Bullet Kin from taking longer to despawn at higher frame rates
 - Added "Flak Bullets Fix", preventing multiple copies of Flak Bullets causing projectiles to duplicate infinitely
+- Added "Scattershot Fix", preventing extra scattershot projectiles from consuming ammo when firing charge weapons (e.g., Crescent Crossbow)
 - Completely disabled "Optimize GUI Mouse Events" option for now, as it causes issues with vanilla menus
 
 ## 1.5.3 (2025-06-20)